requirements.txt
amqp==5.0.6
appdirs==1.4.3
asgiref==3.5.2
astroid==2.5.6
autopep8==1.5.6
Babel==2.9.0
billiard==3.6.4.0
Brotli==1.0.9
CacheControl==0.12.6
cachetools==4.2.4
celery==5.0.5
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
cssselect2==0.4.1
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
Django==3.2
django-crispy-forms==1.11.2
django-embed-video==1.4.1
django-environ==0.8.1
django-etc==1.3.0
django-filter==2.4.0
django-hitcount==1.3.3
django-map-widgets==0.3.2
django-money==1.3.1
django-multiselectfield==0.1.12
django-phonenumber-field==5.0.0
django-pwa==1.0.10
django-taggit==2.0.0
django-taggit-serializer==0.1.7
django-tinymce==3.4.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
fcm-django==1.0.5
firebase-admin==5.1.0
fonttools==4.28.3
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.42.0
grpcio-status==1.42.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.2
idna==2.8
ipaddr==2.2.0
isort==5.8.0
kombu==5.0.2
lazy-object-proxy==1.6.0
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
phonenumbers==8.12.22
Pillow==9.2.0
platformdirs==2.4.0
progress==1.5
prompt-toolkit==3.0.18
proto-plus==1.19.8
protobuf==3.19.1
psycopg2==2.8.6
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydyf==0.2.0
PyJWT==2.0.1
pylint==2.8.2
pyparsing==2.4.6
pyphen==0.11.0
python-http-client==3.3.2
python-memcached==1.59
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.8
sendgrid==6.6.0
six==1.14.0
slack-sdk==3.13.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
tinycss2==1.1.1
toml==0.10.2
twilio==7.3.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
weasyprint==56.0
webencodings==0.5.1
wrapt==1.12.1
zopfli==0.1.9
